Refactor database connection and remove city service
Removed the files related to the CityService and Databases. This includes the SQL functions embedded in the CityService. These files are extracted out, leaving only the function to connect to the Postgres database. Additionally, updated the application.yaml to include Postgres URL and added codeStyleConfig.xml file which tweaks the style options for the project.